Revert " Fix some Wconversion 6/n" (#2467)
* Revert " Fix some Wconversion 6/n (#2338)"

This reverts commit f8d00e2215d13a7b303517c2354284a3609f9790 which caused PT promotion blocker https://jira.habana-labs.com/browse/

+++ hpu_ops/backend/bincount_gen.cpp
@@ -39,7 +39,7 @@ void BinCount::AddNode(synapse_helpers::graph& graph, const at::Stack& stack) {
 
   StackGetter stackGetter(this, stack, "Bincount::AddNode");
   auto self = stackGetter.getNextInput<TensorsPair>();
-  auto length = stackGetter.getNextInput<long>();
+  auto length = stackGetter.getNextInput<int32_t>();
   auto weights = stackGetter.getNextInput<std::optional<TensorsPair>>();
 
   ns_BinCountKernel::Params params{
